permit lazyserviceidentifier for multiinject
