+++ src/annotation/multi_inject.ts
@@ -1,10 +1,6 @@
 import * as METADATA_KEY from "../constants/metadata_keys";
-import { interfaces } from "../interfaces/interfaces";
-import { Metadata } from "../planning/metadata";
-import { createTaggedDecorator } from "./decorator_utils";
+import { injectBase } from "./inject_base";
 
-function multiInject(serviceIdentifier: interfaces.ServiceIdentifier<any>) {
-  return createTaggedDecorator(new Metadata(METADATA_KEY.MULTI_INJECT_TAG, serviceIdentifier));
-}
+const multiInject = injectBase(METADATA_KEY.MULTI_INJECT_TAG);
 
 export { multiInject };
